changed timer from 64 to 32 bit, added alpha beta filert

+++ app/Timing/inc/Timing.hpp
@@ -8,8 +8,8 @@ namespace TIMING
 
 class Ticker{
 private:
-  uint64_t tick_millis;
-  uint64_t tick_micros;
+  uint32_t tick_millis;
+  uint32_t tick_micros;
 public:
 
   /// @brief Construct a new Ticker object
@@ -22,11 +22,11 @@ public:
 
   /// @brief Get current time in microseconds
   /// @return  current time in microseconds [us]
-  uint64_t get_micros();
+  uint32_t get_micros();
 
   /// @brief get time in milliseconds
   /// @return current time in milliseconds [ms]
-  uint64_t get_millis() const;
+  uint32_t get_millis() const;
 
   /// @brief  get time in seconds with microsecond resolution
   /// @return  current time in seconds [s]
@@ -37,8 +37,8 @@ class Timing
 {
 private:
   Ticker &ticker;
-  uint64_t period;
-  uint64_t last_time;
+  uint32_t period;
+  uint32_t last_time;
   bool repeat;
 public:
   /// @brief Construct a new Timing object
@@ -48,7 +48,7 @@ public:
   /// @brief Set the behaviour of the timer
   /// @param period period of the timer in microseconds [us]
   /// @param repeat if the timer should repeat
-  void set_behaviour(uint64_t period, bool repeat=true);
+  void set_behaviour(uint32_t period, bool repeat=true);
 
   /// @brief Check if the timer has triggered
   /// @return true if the timer has triggered
